+++ src/Pages/Home.js
@@ -1,7 +1,10 @@
+import AppNav from "./AppNav";
+
 function Home(props) {
     return (
         <div>
-
+            <AppNav />
+            stuff
         </div>
     )
 }
